Fix the entry widgets for integer/doubledata linked to C variables
Tcl/Tk 8.6 enforces in real time the -textvariable format of variables, so they can be converted to C int/double types.
Lsd relied in the old behavior of enforcing format only when the widget losed focus, simplifying the error management but making the interface less reliable to the user.
Now all the int/double entry boxes use the -validate option to enforce valid numbers and only update the linked variables when user finishes interaction with the widget.