Missing obj and Crash
Brings to the 6.4.1-stable the bugs fixed in 6.4.1 and apparently lost.
The first bug caused the simulation to crash in case not a single
instance of an object remained in the configuration.
The second bug caused the program to crash in after a error_hard
because of the wrong used of exception.