LaTeX Report P
Introduced an additional reporting tool. It creates LaTeX code in the
form of section and paragraph (flexibly) adding labels so as to allow
for intra-document referencing.
To use include the package hyperref and use the command \hr{MyElement}
to generated a linked reference.
Beware of underscores! If a label includes and under_score it is
necessary to manually correct the text adding a slash, as in
under\_score.