LaTeX Report P
Introduced an additional reporting tool. It creates LaTeX code in the
form of section and paragraph (flexibly) adding labels so as to allow
for intra-document referencing.
To use include the package hyperref and use the command \hr{MyElement}
to generated a linked reference.
Beware of underscores! If a label includes an underscore it is
necessary to use \hrf{link}{linkName} and manually correct the link text only
by adding "\string" in front of the underscore(s), as in "under\string_score".
The linkName doesn't need correction, as in "under_score".