Initialdir bug on Macs
When asking to open a configuration the LSD browser offers the latest
directory used, typically from a different model. The bug is likely due
to Tk. It is avoided by removing the option -initialfile in the string.
This fix is limited to correct the command to open a configuration. To
control whether the same bug appears in other uses of tk_getOpenFile