AoR: Adding existing not saved data
In Analysis of Result it is now possible to include degenerate series with the data currently held in the configuration even if they have not been flagged to be saved. The series is composed by one or more data points indicated with the time stamp t equal to the time of computation/update or 0 in case of a never changed parameters.
The new function is activated using the shortcut Control-b from the main AoR window besides an additional option in the Add Series list.